Extensibility: Add binder for events, split logic by sync primitives
